qinvestigate: readline support, start replacing shell
Readline and history is available on the current shell

A replacement shell that automatically parses the script for new
functions and features and makes them available for the user is in
preparation. Part of the underlying architecture (the run_command
function and the parser of the available commands) is included and
available as an experimental feature for the curious.