send_orca_job can deal with (some) parameters requiring end
For sections like "%coords" or "%basis" we can now deal with some of the
keywords which require an 'end' as well. Thanks to @mth91 for pointing
the problem out.

If you discover keywords which require and end and which the script does
not yet recognise, please add them to the 'subsection_keywords' list
via a PR.